requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
cerebras-cloud-sdk>=1.0.0
tavily-python>=0.3.0
mem0ai>=0.1.0
fastapi>=0.111.0
uvicorn>=0.30.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
